requirements.txt
alabaster==0.7.7
anaconda-client==1.4.0
anaconda-navigator==1.1.0
argcomplete==1.0.0
astropy==1.1.2
Babel==2.2.0
beautifulsoup4==4.4.1
bitarray==0.8.1
blaze==0.9.1
bokeh==0.11.1
boto==2.39.0
Bottleneck==1.0.0
cffi==1.5.2
chest==0.2.3
cloudpickle==0.1.1
clyent==1.2.1
colorama==0.3.7
comtypes==1.1.2
conda==4.0.5
conda-build==1.20.0
conda-env==2.4.5
conda-manager==0.3.1
configobj==5.0.6
cryptography==1.3
cycler==0.10.0
Cython==0.23.4
cytoolz==0.7.5
dask==0.8.1
datashape==0.5.1
decorator==4.0.9
dill==0.2.4
docutils==0.12
dynd===c328ab7
et-xmlfile==1.0.1
fastcache==1.0.2
Flask==0.10.1
Flask-Cors==2.1.2
gevent==1.1.0
greenlet==0.4.9
h5py==2.5.0
HeapDict==1.0.0
idna==2.0
ipykernel==4.3.1
ipython==4.1.2
ipython-genutils==0.1.0
ipywidgets==4.1.1
itsdangerous==0.24
jdcal==1.2
jedi==0.9.0
Jinja2==2.8
jsonschema==2.4.0
jupyter==1.0.0
jupyter-client==4.2.2
jupyter-console==4.1.1
jupyter-core==4.1.0
llvmlite==0.9.0
locket==0.2.0
lxml==3.6.0
MarkupSafe==0.23
matplotlib==1.5.1
menuinst==1.3.2
mistune==0.7.2
mpmath==0.19
multipledispatch==0.4.8
nbconvert==4.1.0
nbformat==4.0.1
networkx==1.11
nltk==3.2
nose==1.3.7
notebook==4.1.0
numba==0.24.0
numexpr==2.5
numpy==1.10.4
odo==0.4.2
openpyxl==2.3.2
pandas==0.18.0
partd==0.3.2
path.py==0.0.0
patsy==0.4.0
pep8==1.7.0
pickleshare==0.5
Pillow==3.1.1
ply==3.8
psutil==4.1.0
py==1.4.31
pyasn1==0.1.9
pycosat==0.6.1
pycparser==2.14
pycrypto==2.6.1
pyflakes==1.1.0
Pygments==2.1.1
pymzml==0.7.7
pyOpenSSL==0.15.1
pyparsing==2.0.3
pyreadline==2.1
pytest==2.8.5
python-dateutil==2.5.1
pytz==2016.2
pywin32==220
PyYAML==3.11
pyzmq==15.2.0
QtAwesome==0.3.2
qtconsole==4.2.0
QtPy==1.0
requests==2.9.1
rope-py3k==0.9.4.post1
scikit-image==0.12.3
scikit-learn==0.17.1
scipy==0.17.0
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.10.0
snowballstemmer==1.2.1
sockjs-tornado==1.0.1
Sphinx==1.3.1
sphinx-rtd-theme==0.1.9
spyder==2.3.8
SQLAlchemy==1.0.12
statsmodels==0.6.1
sympy==1.0
tables==3.2.2
toolz==0.7.4
tornado==4.3
traitlets==4.2.1
unicodecsv==0.14.1
Werkzeug==0.11.4
xlrd==0.9.4
XlsxWriter==0.8.4
xlwings==0.7.0
xlwt==1.0.0
